server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "tensorflow",
    "librosa",
    "numpy",
    "scipy",
    "python-multipart",
    "sqlalchemy",
    "pymysql",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-dotenv",
    "pydantic[email]",
    "pydantic-ai[openai]",
    "openai",
]
